+++ libs/common/src/lib/config.ts
@@ -108,6 +108,7 @@ export const DEFAULT_PROCESSOR_PORTFOLIO_SNAPSHOT_COMPUTATION_TIMEOUT =
   ms('30 seconds');
 
 export const DEFAULT_REDACTED_PATHS = [
+  'account.comment',
   'accounts[*].balance',
   'accounts[*].balanceInBaseCurrency',
   'accounts[*].comment',
@@ -126,6 +127,8 @@ export const DEFAULT_REDACTED_PATHS = [
   'activities[*].quantity',
   'activities[*].value',
   'activities[*].valueInBaseCurrency',
+  'assetProfile.symbolMapping',
+  'assetProfile.watchedByCount',
   'balance',
   'balanceInBaseCurrency',
   'balances[*].account.comment',
@@ -133,6 +136,8 @@ export const DEFAULT_REDACTED_PATHS = [
   'balances[*].valueInBaseCurrency',
   'comment',
   'dividendInBaseCurrency',
+  'fee',
+  'feeInAssetProfileCurrency',
   'feeInBaseCurrency',
   'grossPerformance',
   'grossPerformanceWithCurrencyEffect',
@@ -248,6 +253,9 @@ export const HEADER_KEY_IMPERSONATION = 'Impersonation-Id';
 export const HEADER_KEY_TIMEZONE = 'Timezone';
 export const HEADER_KEY_TOKEN = 'Authorization';
 export const HEADER_KEY_SKIP_INTERCEPTOR = 'X-Skip-Interceptor';
+
+export const HTTP_RESPONSE_MESSAGE_IMPERSONATION_UNRESOLVED =
+  'The impersonation identifier cannot be resolved';
 
 export const MAX_TOP_HOLDINGS = 50;