+++ CHANGELOG.md
@@ -10,14 +10,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed the holdings in the portfolio endpoints from a map keyed by the symbol to an array
-- Improved the performance of the portfolio snapshot calculation by indexing the activities
-- Upgraded `angular` from version `21.2.19` to `22.1.2`
-- Upgraded `ngx-device-detector` from version `11.0.0` to `12.0.0`
-- Upgraded `ngx-markdown` from version `21.2.0` to `22.0.0`
 
 ### Fixed
 
 - Fixed the portfolio calculation for holdings with the same symbol from different data sources
+
+## 3.55.0 - 2026-08-19
+
+### Added
+
+- Added the write scopes to the access
+- Added support to merge an asset profile into an existing asset profile in the asset profile dialog of the admin control panel (experimental)
+
+### Changed
+
+- Improved the usability of the account selectors in the transfer cash balance dialog
+- Improved the performance of the portfolio snapshot calculation by indexing the activities
+- Improved the language localization for German (`de`)
+- Improved the language localization for Spanish (`es`)
+- Updated `angular-developer` skills
+- Upgraded `angular` from version `21.2.19` to `22.1.2`
+- Upgraded `ngx-device-detector` from version `11.0.0` to `12.0.0`
+- Upgraded `ngx-markdown` from version `21.2.0` to `22.0.0`
 
 ## 3.54.0 - 2026-08-18